Move a bunch more code out of QgisApp
