+++ src/app/qgisapp.cpp
@@ -98,10 +98,7 @@
 #include "qgsvectorlayersavestyledialog.h"
 #include "maptools/qgsappmaptools.h"
 #include "qgsexpressioncontextutils.h"
-#include "qgsprovidersublayerdetails.h"
-#include "qgsproviderutils.h"
-#include "qgsprovidersublayersdialog.h"
-#include "qgsmaplayerfactory.h"
+
 #include "qgsbrowserwidget.h"
 #include "annotations/qgsannotationitempropertieswidget.h"
 #include "qgsmaptoolmodifyannotation.h"
@@ -128,6 +125,8 @@
 #include "elevation/qgselevationprofilewidget.h"
 
 #include "layers/qgsapplayerhandling.h"
+
+#include "canvas/qgscanvasrefreshblocker.h"
 
 #ifdef HAVE_3D
 #include "qgs3d.h"
@@ -420,7 +419,6 @@ Q_GUI_EXPORT extern int qt_defaultDpiX();
 #include "devtools/querylogger/qgsqueryloggerwidgetfactory.h"
 #include "devtools/profiler/qgsprofilerwidgetfactory.h"
 #include "qgsabstractdatabaseproviderconnection.h"
-#include "qgszipitem.h"
 
 #include "browser/qgsinbuiltdataitemproviders.h"
 
@@ -2258,7 +2256,7 @@ void QgisApp::handleDropUriList( const QgsMimeDataUtils::UriList &lst )
 
     if ( u.layerType == QLatin1String( "collection" ) )
     {
-      openLayer( uri, true );
+      QgsAppLayerHandling::openLayer( uri, true );
     }
     else if ( u.layerType == QLatin1String( "vector" ) )
     {
@@ -2647,7 +2645,7 @@ void QgisApp::dataSourceManager( const QString &pageName )
     connect( mDataSourceManagerDialog, &QgsDataSourceManagerDialog::connectionsChanged, this, &QgisApp::connectionsChanged );
     connect( mDataSourceManagerDialog, SIGNAL( addRasterLayer( QString const &, QString const &, QString const & ) ),
              this, SLOT( addRasterLayer( QString const &, QString const &, QString const & ) ) );
-    connect( mDataSourceManagerDialog, &QgsDataSourceManagerDialog::addRasterLayers, this, [ = ]( const QStringList & layersList ) { addRasterLayers( layersList ); } );
+    connect( mDataSourceManagerDialog, &QgsDataSourceManagerDialog::addRasterLayers, this, [ = ]( const QStringList & layersList ) { QgsAppLayerHandling::addRasterLayers( layersList ); } );
     connect( mDataSourceManagerDialog, SIGNAL( addVectorLayer( QString const &, QString const &, QString const & ) ),
              this, SLOT( addVectorLayer( QString const &, QString const &, QString const & ) ) );
     connect( mDataSourceManagerDialog, SIGNAL( addVectorLayers( QStringList const &, QString const &, QString const & ) ),
@@ -5600,334 +5598,12 @@ QString QgisApp::crsAndFormatAdjustedLayerUri( const QString &uri, const QString
 
 bool QgisApp::addVectorLayers( const QStringList &layerQStringList, const QString &enc, const QString &dataSourceType )
 {
-  return addVectorLayersPrivate( layerQStringList, enc, dataSourceType );
-}
-
-bool QgisApp::addVectorLayersPrivate( const QStringList &layers, const QString &enc, const QString &dataSourceType, const bool guiWarning )
-{
-  //note: this method ONLY supports vector layers from the OGR provider!
-
-  QgsCanvasRefreshBlocker refreshBlocker;
-
-  QList<QgsMapLayer *> layersToAdd;
-  QList<QgsMapLayer *> addedLayers;
-  QgsSettings settings;
-  bool userAskedToAddLayers = false;
-
-  for ( const QString &layerUri : layers )
-  {
-    const QString uri = layerUri.trimmed();
-    QString baseName;
-    if ( dataSourceType == QLatin1String( "file" ) )
-    {
-      QString srcWithoutLayername( uri );
-      int posPipe = srcWithoutLayername.indexOf( '|' );
-      if ( posPipe >= 0 )
-        srcWithoutLayername.resize( posPipe );
-      baseName = QgsProviderUtils::suggestLayerNameFromFilePath( srcWithoutLayername );
-
-      // if needed prompt for zipitem layers
-      QString vsiPrefix = QgsZipItem::vsiPrefix( uri );
-      if ( ! uri.startsWith( QLatin1String( "/vsi" ), Qt::CaseInsensitive ) &&
-           ( vsiPrefix == QLatin1String( "/vsizip/" ) || vsiPrefix == QLatin1String( "/vsitar/" ) ) )
-      {
-        if ( askUserForZipItemLayers( uri, { QgsMapLayerType::VectorLayer } ) )
-          continue;
-      }
-    }
-    else if ( dataSourceType == QLatin1String( "database" ) )
-    {
-      // Try to extract the database name and use it as base name
-      // sublayers names (if any) will be appended to the layer name
-      const QVariantMap parts( QgsProviderRegistry::instance()->decodeUri( QStringLiteral( "ogr" ), uri ) );
-      if ( parts.value( QStringLiteral( "databaseName" ) ).isValid() )
-        baseName = parts.value( QStringLiteral( "databaseName" ) ).toString();
-      else
-        baseName = uri;
-    }
-    else //directory //protocol
-    {
-      baseName = QgsProviderUtils::suggestLayerNameFromFilePath( uri );
-    }
-
-    if ( settings.value( QStringLiteral( "qgis/formatLayerName" ), false ).toBool() )
-    {
-      baseName = QgsMapLayer::formatLayerName( baseName );
-    }
-
-    QgsDebugMsgLevel( "completeBaseName: " + baseName, 2 );
-    const bool isVsiCurl { uri.startsWith( QLatin1String( "/vsicurl" ), Qt::CaseInsensitive ) };
-    const auto scheme { QUrl( uri ).scheme() };
-    const bool isRemoteUrl { scheme.startsWith( QLatin1String( "http" ) ) || scheme == QLatin1String( "ftp" ) };
-
-    std::unique_ptr< QgsTemporaryCursorOverride > cursorOverride;
-    if ( isVsiCurl || isRemoteUrl )
-    {
-      cursorOverride = std::make_unique< QgsTemporaryCursorOverride >( Qt::WaitCursor );
-      visibleMessageBar()->pushInfo( tr( "Remote layer" ), tr( "loading %1, please wait …" ).arg( uri ) );
-      qApp->processEvents();
-    }
-
-    QList< QgsProviderSublayerDetails > sublayers = QgsProviderRegistry::instance()->providerMetadata( QStringLiteral( "ogr" ) )->querySublayers( uri, Qgis::SublayerQueryFlag::IncludeSystemTables );
-    // filter out non-vector sublayers
-    sublayers.erase( std::remove_if( sublayers.begin(), sublayers.end(), []( const QgsProviderSublayerDetails & sublayer )
-    {
-      return sublayer.type() != QgsMapLayerType::VectorLayer;
-    } ), sublayers.end() );
-
-    cursorOverride.reset();
-
-    const QVariantMap uriParts = QgsProviderRegistry::instance()->decodeUri( QStringLiteral( "ogr" ), uri );
-    const QString path = uriParts.value( QStringLiteral( "path" ) ).toString();
-
-    if ( !sublayers.empty() )
-    {
-      userAskedToAddLayers = true;
-
-      const bool detailsAreIncomplete = QgsProviderUtils::sublayerDetailsAreIncomplete( sublayers, QgsProviderUtils::SublayerCompletenessFlag::IgnoreUnknownFeatureCount );
-      const bool singleSublayerOnly = sublayers.size() == 1;
-      QString groupName;
-
-      if ( !singleSublayerOnly || detailsAreIncomplete )
-      {
-        // ask user for sublayers (unless user settings dictate otherwise!)
-        switch ( shouldAskUserForSublayers( sublayers ) )
-        {
-          case SublayerHandling::AskUser:
-          {
-            // prompt user for sublayers
-            QgsProviderSublayersDialog dlg( uri, path, sublayers, {QgsMapLayerType::VectorLayer}, this );
-
-            if ( dlg.exec() )
-              sublayers = dlg.selectedLayers();
-            else
-              sublayers.clear(); // dialog was canceled, so don't add any sublayers
-            groupName = dlg.groupName();
-            break;
-          }
-
-          case SublayerHandling::LoadAll:
-          {
-            if ( detailsAreIncomplete )
-            {
-              // requery sublayers, resolving geometry types
-              sublayers = QgsProviderRegistry::instance()->querySublayers( uri, Qgis::SublayerQueryFlag::ResolveGeometryType );
-              // filter out non-vector sublayers
-              sublayers.erase( std::remove_if( sublayers.begin(), sublayers.end(), []( const QgsProviderSublayerDetails & sublayer )
-              {
-                return sublayer.type() != QgsMapLayerType::VectorLayer;
-              } ), sublayers.end() );
-            }
-            break;
-          }
-
-          case SublayerHandling::AbortLoading:
-            sublayers.clear(); // don't add any sublayers
-            break;
-        };
-      }
-      else if ( detailsAreIncomplete )
-      {
-        // requery sublayers, resolving geometry types
-        sublayers = QgsProviderRegistry::instance()->querySublayers( uri, Qgis::SublayerQueryFlag::ResolveGeometryType );
-        // filter out non-vector sublayers
-        sublayers.erase( std::remove_if( sublayers.begin(), sublayers.end(), []( const QgsProviderSublayerDetails & sublayer )
-        {
-          return sublayer.type() != QgsMapLayerType::VectorLayer;
-        } ), sublayers.end() );
-      }
-
-      // now add sublayers
-      if ( !sublayers.empty() )
-      {
-        addedLayers << addSublayers( sublayers, baseName, groupName );
-      }
-
-    }
-    else
-    {
-      QString msg = tr( "%1 is not a valid or recognized data source." ).arg( uri );
-      // If the failed layer was a vsicurl type, give the user a chance to try the normal download.
-      if ( isVsiCurl &&
-           QMessageBox::question( this, tr( "Invalid Data Source" ),
-                                  tr( "Download with \"Protocol\" source type has failed, do you want to try the \"File\" source type?" ) ) == QMessageBox::Yes )
-      {
-        QString fileUri = uri;
-        fileUri.replace( QLatin1String( "/vsicurl/" ), " " );
-        return addVectorLayersPrivate( QStringList() << fileUri, enc, dataSourceType, guiWarning );
-      }
-      else if ( guiWarning )
-      {
-        visibleMessageBar()->pushMessage( tr( "Invalid Data Source" ), msg, Qgis::MessageLevel::Critical );
-      }
-    }
-  }
-
-  // make sure at least one layer was successfully added
-  if ( layersToAdd.isEmpty() )
-  {
-    // we also return true if we asked the user for sublayers, but they choose none. In this case nothing
-    // went wrong, so we shouldn't return false and cause GUI warnings to appear
-    return userAskedToAddLayers || !addedLayers.isEmpty();
-  }
-
-  // Register this layer with the layers registry
-  QgsProject::instance()->addMapLayers( layersToAdd );
-  for ( QgsMapLayer *l : std::as_const( layersToAdd ) )
-  {
-    if ( !enc.isEmpty() )
-    {
-      if ( QgsVectorLayer *vl = qobject_cast< QgsVectorLayer * >( l ) )
-        vl->setProviderEncoding( enc );
-    }
-
-    askUserForDatumTransform( l->crs(), QgsProject::instance()->crs(), l );
-    QgsAppLayerHandling::postProcessAddedLayer( l );
-  }
-  activateDeactivateLayerRelatedActions( activeLayer() );
-
-  return true;
+  return QgsAppLayerHandling::addVectorLayers( layerQStringList, enc, dataSourceType );
 }
 
 QgsMeshLayer *QgisApp::addMeshLayer( const QString &url, const QString &baseName, const QString &providerKey )
 {
-  return addLayerPrivate< QgsMeshLayer >( QgsMapLayerType::MeshLayer, url, baseName, providerKey, true );
-}
-
-QList< QgsMapLayer * > QgisApp::addSublayers( const QList<QgsProviderSublayerDetails> &layers, const QString &baseName, const QString &groupName )
-{
-  QgsLayerTreeGroup *group = nullptr;
-  if ( !groupName.isEmpty() )
-  {
-    int index { 0 };
-    QgsLayerTreeNode *currentNode { mLayerTreeView->currentNode() };
-    if ( currentNode && currentNode->parent() )
-    {
-      if ( QgsLayerTree::isGroup( currentNode ) )
-      {
-        group = qobject_cast<QgsLayerTreeGroup *>( currentNode )->insertGroup( 0, groupName );
-      }
-      else if ( QgsLayerTree::isLayer( currentNode ) )
-      {
-        const QList<QgsLayerTreeNode *> currentNodeSiblings { currentNode->parent()->children() };
-        int nodeIdx { 0 };
-        for ( const QgsLayerTreeNode *child : std::as_const( currentNodeSiblings ) )
-        {
-          nodeIdx++;
-          if ( child == currentNode )
-          {
-            index = nodeIdx;
-            break;
-          }
-        }
-        group = qobject_cast<QgsLayerTreeGroup *>( currentNode->parent() )->insertGroup( index, groupName );
-      }
-      else
-      {
-        group = QgsProject::instance()->layerTreeRoot()->insertGroup( 0, groupName );
-      }
-    }
-    else
-    {
-      group = QgsProject::instance()->layerTreeRoot()->insertGroup( 0, groupName );
-    }
-  }
-
-  QgsSettings settings;
-  const bool formatLayerNames = settings.value( QStringLiteral( "qgis/formatLayerName" ), false ).toBool();
-
-  // if we aren't adding to a group, we need to add the layers in reverse order so that they maintain the correct
-  // order in the layer tree!
-  QList<QgsProviderSublayerDetails> sortedLayers = layers;
-  if ( groupName.isEmpty() )
-  {
-    std::reverse( sortedLayers.begin(), sortedLayers.end() );
-  }
-
-  QList< QgsMapLayer * > result;
-  result.reserve( sortedLayers.size() );
-
-  for ( const QgsProviderSublayerDetails &sublayer : std::as_const( sortedLayers ) )
-  {
-    QgsProviderSublayerDetails::LayerOptions options( QgsProject::instance()->transformContext() );
-    options.loadDefaultStyle = false;
-
-    std::unique_ptr<QgsMapLayer> layer( sublayer.toLayer( options ) );
-    if ( !layer )
-      continue;
-
-    QgsMapLayer *ml = layer.get();
-    // if we aren't adding to a group, then we're iterating the layers in the reverse order
-    // so account for that in the returned list of layers
-    if ( groupName.isEmpty() )
-      result.insert( 0, ml );
-    else
-      result << ml;
-
-    QString layerName = layer->name();
-    if ( formatLayerNames )
-    {
-      layerName = QgsMapLayer::formatLayerName( layerName );
-    }
-
-    const bool projectWasEmpty = QgsProject::instance()->mapLayers().empty();
-
-    // if user has opted to add sublayers to a group, then we don't need to include the
-    // filename in the layer's name, because the group is already titled with the filename.
-    // But otherwise, we DO include the file name so that users can differentiate the source
-    // when multiple layers are loaded from a GPX file or similar (refs https://github.com/qgis/QGIS/issues/37551)
-    if ( group )
-    {
-      if ( !layerName.isEmpty() )
-        layer->setName( layerName );
-      else if ( !baseName.isEmpty() )
-        layer->setName( baseName );
-      QgsProject::instance()->addMapLayer( layer.release(), false );
-      group->addLayer( ml );
-    }
-    else
-    {
-      if ( layerName != baseName && !layerName.isEmpty() && !baseName.isEmpty() )
-        layer->setName( QStringLiteral( "%1 — %2" ).arg( baseName, layerName ) );
-      else if ( !layerName.isEmpty() )
-        layer->setName( layerName );
-      else if ( !baseName.isEmpty() )
-        layer->setName( baseName );
-      QgsProject::instance()->addMapLayer( layer.release() );
-    }
-
-    // Some of the logic relating to matching a new project's CRS to the first layer added CRS is deferred to happen when the event loop
-    // next runs -- so in those cases we can't assume that the project's CRS has been matched to the actual desired CRS yet.
-    // In these cases we don't need to show the coordinate operation selection choice, so just hardcode an exception in here to avoid that...
-    QgsCoordinateReferenceSystem projectCrsAfterLayerAdd = QgsProject::instance()->crs();
-    const QgsGui::ProjectCrsBehavior projectCrsBehavior = QgsSettings().enumValue( QStringLiteral( "/projections/newProjectCrsBehavior" ),  QgsGui::UseCrsOfFirstLayerAdded, QgsSettings::App );
-    switch ( projectCrsBehavior )
-    {
-      case QgsGui::UseCrsOfFirstLayerAdded:
-      {
-        if ( projectWasEmpty )
-          projectCrsAfterLayerAdd = ml->crs();
-        break;
-      }
-
-      case QgsGui::UsePresetCrs:
-        break;
-    }
-
-    askUserForDatumTransform( ml->crs(), projectCrsAfterLayerAdd, ml );
-    QgsAppLayerHandling::postProcessAddedLayer( ml );
-  }
-
-  if ( group )
-  {
-    // Respect if user don't want the new group of layers visible.
-    QgsSettings settings;
-    const bool newLayersVisible = settings.value( QStringLiteral( "/qgis/new_layers_visible" ), true ).toBool();
-    if ( !newLayersVisible )
-      group->setItemVisibilityCheckedRecursive( newLayersVisible );
-  }
-
-  return result;
+  return QgsAppLayerHandling::addLayerPrivate< QgsMeshLayer >( QgsMapLayerType::MeshLayer, url, baseName, providerKey, true );
 }
 
 QgsVectorTileLayer *QgisApp::addVectorTileLayer( const QString &url, const QString &baseName )
@@ -5937,7 +5613,7 @@ QgsVectorTileLayer *QgisApp::addVectorTileLayer( const QString &url, const QStri
 
 QgsPointCloudLayer *QgisApp::addPointCloudLayer( const QString &url, const QString &baseName, const QString &providerKey )
 {
-  return addPointCloudLayerPrivate( url, baseName, providerKey );
+  return QgsAppLayerHandling::addPointCloudLayer( url, baseName, providerKey );
 }
 
 QgsVectorTileLayer *QgisApp::addVectorTileLayerPrivate( const QString &url, const QString &baseName, const bool guiWarning )
@@ -5976,162 +5652,6 @@ QgsVectorTileLayer *QgisApp::addVectorTileLayerPrivate( const QString &url, cons
   activateDeactivateLayerRelatedActions( activeLayer() );
 
   return layer.release();
-}
-
-QgsPointCloudLayer *QgisApp::addPointCloudLayerPrivate( const QString &uri, const QString &baseName, const QString &providerKey, bool guiWarning )
-{
-  QgsCanvasRefreshBlocker refreshBlocker;
-  QgsSettings settings;
-
-  QString base( baseName );
-
-  if ( settings.value( QStringLiteral( "qgis/formatLayerName" ), false ).toBool() )
-  {
-    base = QgsMapLayer::formatLayerName( base );
-  }
-
-  QgsDebugMsgLevel( "completeBaseName: " + base, 2 );
-
-  // create the layer
-  std::unique_ptr<QgsPointCloudLayer> layer( new QgsPointCloudLayer( uri, base, providerKey ) );
-
-  if ( !layer || !layer->isValid() )
-  {
-    if ( guiWarning )
-    {
-      QString msg = tr( "%1 is not a valid or recognized data source, error: \"%2\"" ).arg( uri ).arg( layer->error().message( QgsErrorMessage::Format::Text ) );
-      visibleMessageBar()->pushMessage( tr( "Invalid Data Source" ), msg, Qgis::MessageLevel::Critical );
-    }
-
-    // since the layer is bad, stomp on it
-    return nullptr;
-  }
-
-  QgsAppLayerHandling::postProcessAddedLayer( layer.get() );
-
-
-  QgsProject::instance()->addMapLayer( layer.get() );
-  activateDeactivateLayerRelatedActions( activeLayer() );
-
-  return layer.release();
-}
-
-bool QgisApp::askUserForZipItemLayers( const QString &path, const QList< QgsMapLayerType > &acceptableTypes )
-{
-  // query sublayers
-  QList< QgsProviderSublayerDetails > sublayers = QgsProviderRegistry::instance()->querySublayers( path, Qgis::SublayerQueryFlag::IncludeSystemTables );
-
-  // filter out non-matching sublayers
-  sublayers.erase( std::remove_if( sublayers.begin(), sublayers.end(), [acceptableTypes]( const QgsProviderSublayerDetails & sublayer )
-  {
-    return !acceptableTypes.empty() && !acceptableTypes.contains( sublayer.type() );
-  } ), sublayers.end() );
-
-  if ( sublayers.empty() )
-    return false;
-
-  const bool detailsAreIncomplete = QgsProviderUtils::sublayerDetailsAreIncomplete( sublayers, QgsProviderUtils::SublayerCompletenessFlag::IgnoreUnknownFeatureCount );
-  const bool singleSublayerOnly = sublayers.size() == 1;
-  QString groupName;
-
-  if ( !singleSublayerOnly || detailsAreIncomplete )
-  {
-    // ask user for sublayers (unless user settings dictate otherwise!)
-    switch ( shouldAskUserForSublayers( sublayers ) )
-    {
-      case SublayerHandling::AskUser:
-      {
-        // prompt user for sublayers
-        QgsProviderSublayersDialog dlg( path, path, sublayers, acceptableTypes, this );
-
-        if ( dlg.exec() )
-          sublayers = dlg.selectedLayers();
-        else
-          sublayers.clear(); // dialog was canceled, so don't add any sublayers
-        groupName = dlg.groupName();
-        break;
-      }
-
-      case SublayerHandling::LoadAll:
-      {
-        if ( detailsAreIncomplete )
-        {
-          // requery sublayers, resolving geometry types
-          sublayers = QgsProviderRegistry::instance()->querySublayers( path, Qgis::SublayerQueryFlag::ResolveGeometryType );
-          sublayers.erase( std::remove_if( sublayers.begin(), sublayers.end(), [acceptableTypes]( const QgsProviderSublayerDetails & sublayer )
-          {
-            return !acceptableTypes.empty() && !acceptableTypes.contains( sublayer.type() );
-          } ), sublayers.end() );
-        }
-        break;
-      }
-
-      case SublayerHandling::AbortLoading:
-        sublayers.clear(); // don't add any sublayers
-        break;
-    };
-  }
-  else if ( detailsAreIncomplete )
-  {
-    // requery sublayers, resolving geometry types
-    sublayers = QgsProviderRegistry::instance()->querySublayers( path, Qgis::SublayerQueryFlag::ResolveGeometryType );
-    sublayers.erase( std::remove_if( sublayers.begin(), sublayers.end(), [acceptableTypes]( const QgsProviderSublayerDetails & sublayer )
-    {
-      return !acceptableTypes.empty() && !acceptableTypes.contains( sublayer.type() );
-    } ), sublayers.end() );
-  }
-
-  // now add sublayers
-  if ( !sublayers.empty() )
-  {
-    QgsCanvasRefreshBlocker refreshBlocker;
-    QgsSettings settings;
-
-    QString base = QgsProviderUtils::suggestLayerNameFromFilePath( path );
-    if ( settings.value( QStringLiteral( "qgis/formatLayerName" ), false ).toBool() )
-    {
-      base = QgsMapLayer::formatLayerName( base );
-    }
-
-    addSublayers( sublayers, base, groupName );
-    activateDeactivateLayerRelatedActions( activeLayer() );
-  }
-
-  return true;
-}
-
-QgisApp::SublayerHandling QgisApp::shouldAskUserForSublayers( const QList<QgsProviderSublayerDetails> &layers, bool hasNonLayerItems ) const
-{
-  if ( hasNonLayerItems )
-    return SublayerHandling::AskUser;
-
-  QgsSettings settings;
-  const Qgis::SublayerPromptMode promptLayers = settings.enumValue( QStringLiteral( "qgis/promptForSublayers" ), Qgis::SublayerPromptMode::AlwaysAsk );
-
-  switch ( promptLayers )
-  {
-    case Qgis::SublayerPromptMode::AlwaysAsk:
-      return SublayerHandling::AskUser;
-
-    case Qgis::SublayerPromptMode::AskExcludingRasterBands:
-    {
-      // if any non-raster layers are found, we ask the user. Otherwise we load all
-      for ( const QgsProviderSublayerDetails &sublayer : layers )
-      {
-        if ( sublayer.type() != QgsMapLayerType::RasterLayer )
-          return SublayerHandling::AskUser;
-      }
-      return SublayerHandling::LoadAll;
-    }
-
-    case Qgis::SublayerPromptMode::NeverAskSkip:
-      return SublayerHandling::AbortLoading;
-
-    case Qgis::SublayerPromptMode::NeverAskLoadAll:
-      return SublayerHandling::LoadAll;
-  }
-
-  return SublayerHandling::AskUser;
 }
 
 void QgisApp::addDatabaseLayers( QStringList const &layerPathList, QString const &providerKey )
@@ -7540,229 +7060,6 @@ void QgisApp::openProject( const QString &fileName )
   }
 }
 
-bool QgisApp::openLayer( const QString &fileName, bool allowInteractive )
-{
-  bool ok = false;
-  const QFileInfo fileInfo( fileName );
-
-  // highest priority = delegate to provider registry to handle
-  const QList< QgsProviderRegistry::ProviderCandidateDetails > candidateProviders = QgsProviderRegistry::instance()->preferredProvidersForUri( fileName );
-  if ( candidateProviders.size() == 1 && candidateProviders.at( 0 ).layerTypes().size() == 1 )
-  {
-    // one good candidate provider and possible layer type -- that makes things nice and easy!
-    switch ( candidateProviders.at( 0 ).layerTypes().at( 0 ) )
-    {
-      case QgsMapLayerType::VectorLayer:
-      case QgsMapLayerType::RasterLayer:
-      case QgsMapLayerType::MeshLayer:
-      case QgsMapLayerType::AnnotationLayer:
-      case QgsMapLayerType::PluginLayer:
-      case QgsMapLayerType::VectorTileLayer:
-      case QgsMapLayerType::GroupLayer:
-        // not supported here yet!
-        break;
-
-      case QgsMapLayerType::PointCloudLayer:
-        ok = static_cast< bool >( addPointCloudLayerPrivate( fileName, fileInfo.completeBaseName(), candidateProviders.at( 0 ).metadata()->key(), false ) );
-        break;
-    }
-  }
-
-  if ( ok )
-    return true;
-
-  CPLPushErrorHandler( CPLQuietErrorHandler );
-
-  // if needed prompt for zipitem layers
-  QString vsiPrefix = QgsZipItem::vsiPrefix( fileName );
-  if ( vsiPrefix == QLatin1String( "/vsizip/" ) || vsiPrefix == QLatin1String( "/vsitar/" ) )
-  {
-    if ( askUserForZipItemLayers( fileName, {} ) )
-    {
-      CPLPopErrorHandler();
-      return true;
-    }
-  }
-
-  if ( fileName.endsWith( QStringLiteral( ".mbtiles" ), Qt::CaseInsensitive ) )
-  {
-    QgsMbTiles reader( fileName );
-    if ( reader.open() )
-    {
-      if ( reader.metadataValue( "format" ) == QLatin1String( "pbf" ) )
-      {
-        // these are vector tiles
-        QUrlQuery uq;
-        uq.addQueryItem( QStringLiteral( "type" ), QStringLiteral( "mbtiles" ) );
-        uq.addQueryItem( QStringLiteral( "url" ), fileName );
-        const QgsVectorTileLayer::LayerOptions options( QgsProject::instance()->transformContext() );
-        std::unique_ptr<QgsVectorTileLayer> vtLayer( new QgsVectorTileLayer( uq.toString(), fileInfo.completeBaseName(), options ) );
-        if ( vtLayer->isValid() )
-        {
-          QgsProject::instance()->addMapLayer( vtLayer.release() );
-          return true;
-        }
-      }
-      else // raster tiles
-      {
-        // prefer to use WMS provider's implementation to open MBTiles rasters
-        QUrlQuery uq;
-        uq.addQueryItem( QStringLiteral( "type" ), QStringLiteral( "mbtiles" ) );
-        uq.addQueryItem( QStringLiteral( "url" ), QUrl::fromLocalFile( fileName ).toString() );
-        if ( addRasterLayer( uq.toString(), fileInfo.completeBaseName(), QStringLiteral( "wms" ) ) )
-          return true;
-      }
-    }
-  }
-  else if ( fileName.endsWith( QStringLiteral( ".vtpk" ), Qt::CaseInsensitive ) )
-  {
-    // these are vector tiles
-    QUrlQuery uq;
-    uq.addQueryItem( QStringLiteral( "type" ), QStringLiteral( "vtpk" ) );
-    uq.addQueryItem( QStringLiteral( "url" ), fileName );
-    const QgsVectorTileLayer::LayerOptions options( QgsProject::instance()->transformContext() );
-    std::unique_ptr<QgsVectorTileLayer> vtLayer( new QgsVectorTileLayer( uq.toString(), fileInfo.completeBaseName(), options ) );
-    if ( vtLayer->isValid() )
-    {
-      QgsAppLayerHandling::postProcessAddedLayer( vtLayer.get() );
-      QgsProject::instance()->addMapLayer( vtLayer.release() );
-      return true;
-    }
-  }
-
-  QList< QgsProviderSublayerModel::NonLayerItem > nonLayerItems;
-  if ( QgsProjectStorage *ps = QgsApplication::projectStorageRegistry()->projectStorageFromUri( fileName ) )
-  {
-    const QStringList projects = ps->listProjects( fileName );
-    for ( const QString &project : projects )
-    {
-      QgsProviderSublayerModel::NonLayerItem projectItem;
-      projectItem.setType( QStringLiteral( "project" ) );
-      projectItem.setName( project );
-      projectItem.setUri( QStringLiteral( "%1://%2?projectName=%3" ).arg( ps->type(), fileName, project ) );
-      projectItem.setIcon( QgsApplication::getThemeIcon( QStringLiteral( "/mIconQgsProjectFile.svg" ) ) );
-      nonLayerItems << projectItem;
-    }
-  }
-
-  // query sublayers
-  QList< QgsProviderSublayerDetails > sublayers = QgsProviderRegistry::instance()->querySublayers( fileName, Qgis::SublayerQueryFlag::IncludeSystemTables );
-
-  if ( !sublayers.empty() || !nonLayerItems.empty() )
-  {
-    const bool detailsAreIncomplete = QgsProviderUtils::sublayerDetailsAreIncomplete( sublayers, QgsProviderUtils::SublayerCompletenessFlag::IgnoreUnknownFeatureCount );
-    const bool singleSublayerOnly = sublayers.size() == 1;
-    QString groupName;
-
-    if ( allowInteractive && ( !singleSublayerOnly || detailsAreIncomplete || !nonLayerItems.empty() ) )
-    {
-      // ask user for sublayers (unless user settings dictate otherwise!)
-      switch ( shouldAskUserForSublayers( sublayers, !nonLayerItems.empty() ) )
-      {
-        case SublayerHandling::AskUser:
-        {
-          // prompt user for sublayers
-          QgsProviderSublayersDialog dlg( fileName, fileName, sublayers, {}, this );
-          dlg.setNonLayerItems( nonLayerItems );
-
-          if ( dlg.exec() )
-          {
-            sublayers = dlg.selectedLayers();
-            nonLayerItems = dlg.selectedNonLayerItems();
-          }
-          else
-          {
-            sublayers.clear(); // dialog was canceled, so don't add any sublayers
-            nonLayerItems.clear();
-          }
-          groupName = dlg.groupName();
-          break;
-        }
-
-        case SublayerHandling::LoadAll:
-        {
-          if ( detailsAreIncomplete )
-          {
-            // requery sublayers, resolving geometry types
-            sublayers = QgsProviderRegistry::instance()->querySublayers( fileName, Qgis::SublayerQueryFlag::ResolveGeometryType );
-          }
-          break;
-        }
-
-        case SublayerHandling::AbortLoading:
-          sublayers.clear(); // don't add any sublayers
-          break;
-      };
-    }
-    else if ( detailsAreIncomplete )
-    {
-      // requery sublayers, resolving geometry types
-      sublayers = QgsProviderRegistry::instance()->querySublayers( fileName, Qgis::SublayerQueryFlag::ResolveGeometryType );
-    }
-
-    ok = true;
-
-    // now add sublayers
-    if ( !sublayers.empty() )
-    {
-      QgsCanvasRefreshBlocker refreshBlocker;
-      QgsSettings settings;
-
-      QString base = QgsProviderUtils::suggestLayerNameFromFilePath( fileName );
-      if ( settings.value( QStringLiteral( "qgis/formatLayerName" ), false ).toBool() )
-      {
-        base = QgsMapLayer::formatLayerName( base );
-      }
-
-      addSublayers( sublayers, base, groupName );
-      activateDeactivateLayerRelatedActions( activeLayer() );
-    }
-    else if ( !nonLayerItems.empty() )
-    {
-      QgsCanvasRefreshBlocker refreshBlocker;
-      if ( checkTasksDependOnProject() )
-        return true;
-
-      // possibly save any pending work before opening a different project
-      if ( checkUnsavedLayerEdits() && checkMemoryLayers() && saveDirty() )
-      {
-        // error handling and reporting is in addProject() function
-        addProject( nonLayerItems.at( 0 ).uri() );
-      }
-      return true;
-    }
-  }
-
-  CPLPopErrorHandler();
-
-  if ( !ok )
-  {
-    // maybe a known file type, which couldn't be opened due to a missing dependency... (eg. las for a non-pdal-enabled build)
-    QgsProviderRegistry::UnusableUriDetails details;
-    if ( QgsProviderRegistry::instance()->handleUnusableUri( fileName, details ) )
-    {
-      ok = true;
-
-      if ( details.detailedWarning.isEmpty() )
-        visibleMessageBar()->pushMessage( QString(), details.warning, Qgis::MessageLevel::Critical );
-      else
-        visibleMessageBar()->pushMessage( QString(), details.warning, details.detailedWarning, Qgis::MessageLevel::Critical );
-    }
-  }
-
-  if ( !ok )
-  {
-    // we have no idea what this file is...
-    QgsMessageLog::logMessage( tr( "Unable to load %1" ).arg( fileName ) );
-
-    const QString msg = tr( "%1 is not a valid or recognized data source." ).arg( fileName );
-    visibleMessageBar()->pushMessage( tr( "Invalid Data Source" ), msg, Qgis::MessageLevel::Critical );
-  }
-
-  return ok;
-}
-
-
 // Open a file specified by a commandline argument, Drop or FileOpen event.
 void QgisApp::openFile( const QString &fileName, const QString &fileTypeHint )
 {
@@ -7788,7 +7085,7 @@ void QgisApp::openFile( const QString &fileName, const QString &fileTypeHint )
   else
   {
     QgsDebugMsgLevel( "Adding " + fileName + " to the map canvas", 2 );
-    openLayer( fileName, true );
+    QgsAppLayerHandling::openLayer( fileName, true );
   }
 }
 
@@ -8993,11 +8290,11 @@ QString QgisApp::saveAsRasterFile( QgsRasterLayer *rasterLayer, const bool defau
     {
       if ( outputFormat == QLatin1String( "GPKG" ) && !outputLayerName.isEmpty() )
       {
-        addRasterLayers( QStringList( QStringLiteral( "GPKG:%1:%2" ).arg( fileName, outputLayerName ) ) );
+        QgsAppLayerHandling::addRasterLayers( QStringList( QStringLiteral( "GPKG:%1:%2" ).arg( fileName, outputLayerName ) ) );
       }
       else
       {
-        addRasterLayers( QStringList( fileName ) );
+        QgsAppLayerHandling::addRasterLayers( QStringList( fileName ) );
       }
     }
     if ( rlWeakPointer )
@@ -13712,131 +13009,7 @@ void QgisApp::show3DMapViewsManager()
 
 QgsVectorLayer *QgisApp::addVectorLayer( const QString &vectorLayerPath, const QString &name, const QString &providerKey )
 {
-  return addLayerPrivate< QgsVectorLayer >( QgsMapLayerType::VectorLayer, vectorLayerPath, name, !providerKey.isEmpty() ? providerKey : QLatin1String( "ogr" ), true );
-}
-
-template<typename T>
-T *QgisApp::addLayerPrivate( QgsMapLayerType type, const QString &uri, const QString &name, const QString &providerKey, bool guiWarnings )
-{
-  QgsSettings settings;
-
-  QgsCanvasRefreshBlocker refreshBlocker;
-
-  QString baseName = settings.value( QStringLiteral( "qgis/formatLayerName" ), false ).toBool() ? QgsMapLayer::formatLayerName( name ) : name;
-
-  // if the layer needs authentication, ensure the master password is set
-  const QRegularExpression rx( "authcfg=([a-z]|[A-Z]|[0-9]){7}" );
-  if ( rx.match( uri ).hasMatch() )
-  {
-    if ( !QgsAuthGuiUtils::isDisabled( messageBar() ) )
-    {
-      QgsApplication::authManager()->setMasterPassword( true );
-    }
-  }
-
-  QVariantMap uriElements = QgsProviderRegistry::instance()->decodeUri( providerKey, uri );
-  QString path = uri;
-  if ( uriElements.contains( QStringLiteral( "path" ) ) )
-  {
-    // run layer path through QgsPathResolver so that all inbuilt paths and other localised paths are correctly expanded
-    path = QgsPathResolver().readPath( uriElements.value( QStringLiteral( "path" ) ).toString() );
-    uriElements[ QStringLiteral( "path" ) ] = path;
-  }
-  // Not all providers implement decodeUri(), so use original uri if uriElements is empty
-  const QString updatedUri = uriElements.isEmpty() ? uri : QgsProviderRegistry::instance()->encodeUri( providerKey, uriElements );
-
-  const bool canQuerySublayers = QgsProviderRegistry::instance()->providerMetadata( providerKey ) &&
-                                 ( QgsProviderRegistry::instance()->providerMetadata( providerKey )->capabilities() & QgsProviderMetadata::QuerySublayers );
-
-  T *result = nullptr;
-  if ( canQuerySublayers )
-  {
-    // query sublayers
-    QList< QgsProviderSublayerDetails > sublayers = QgsProviderRegistry::instance()->providerMetadata( providerKey ) ?
-        QgsProviderRegistry::instance()->providerMetadata( providerKey )->querySublayers( updatedUri, Qgis::SublayerQueryFlag::IncludeSystemTables )
-        : QgsProviderRegistry::instance()->querySublayers( updatedUri );
-
-    // filter out non-matching sublayers
-    sublayers.erase( std::remove_if( sublayers.begin(), sublayers.end(), [type]( const QgsProviderSublayerDetails & sublayer )
-    {
-      return sublayer.type() != type;
-    } ), sublayers.end() );
-
-    if ( sublayers.empty() )
-    {
-      if ( guiWarnings )
-      {
-        QString msg = tr( "%1 is not a valid or recognized data source." ).arg( uri );
-        visibleMessageBar()->pushMessage( tr( "Invalid Data Source" ), msg, Qgis::MessageLevel::Critical );
-      }
-
-      // since the layer is bad, stomp on it
-      return nullptr;
-    }
-    else if ( sublayers.size() > 1 || QgsProviderUtils::sublayerDetailsAreIncomplete( sublayers, QgsProviderUtils::SublayerCompletenessFlag::IgnoreUnknownFeatureCount ) )
-    {
-      // ask user for sublayers (unless user settings dictate otherwise!)
-      switch ( shouldAskUserForSublayers( sublayers ) )
-      {
-        case SublayerHandling::AskUser:
-        {
-          QgsProviderSublayersDialog dlg( updatedUri, path, sublayers, {type}, this );
-          if ( dlg.exec() )
-          {
-            const QList< QgsProviderSublayerDetails > selectedLayers = dlg.selectedLayers();
-            if ( !selectedLayers.isEmpty() )
-            {
-              result = qobject_cast< T * >( addSublayers( selectedLayers, baseName, dlg.groupName() ).value( 0 ) );
-            }
-          }
-          break;
-        }
-        case SublayerHandling::LoadAll:
-        {
-          result = qobject_cast< T * >( addSublayers( sublayers, baseName, QString() ).value( 0 ) );
-          break;
-        }
-        case SublayerHandling::AbortLoading:
-          break;
-      };
-    }
-    else
-    {
-      result = qobject_cast< T * >( addSublayers( sublayers, name, QString() ).value( 0 ) );
-
-      if ( result )
-      {
-        QString base( baseName );
-        if ( settings.value( QStringLiteral( "qgis/formatLayerName" ), false ).toBool() )
-        {
-          base = QgsMapLayer::formatLayerName( base );
-        }
-        result->setName( base );
-      }
-    }
-  }
-  else
-  {
-    QgsMapLayerFactory::LayerOptions options( QgsProject::instance()->transformContext() );
-    options.loadDefaultStyle = false;
-    result = qobject_cast< T * >( QgsMapLayerFactory::createLayer( uri, name, type, options, providerKey ) );
-    if ( result )
-    {
-      QString base( baseName );
-      if ( settings.value( QStringLiteral( "qgis/formatLayerName" ), false ).toBool() )
-      {
-        base = QgsMapLayer::formatLayerName( base );
-      }
-      result->setName( base );
-      QgsProject::instance()->addMapLayer( result );
-
-      askUserForDatumTransform( result->crs(), QgsProject::instance()->crs(), result );
-      QgsAppLayerHandling::postProcessAddedLayer( result );
-    }
-  }
-
-  activateDeactivateLayerRelatedActions( activeLayer() );
-  return result;
+  return QgsAppLayerHandling::addLayerPrivate< QgsVectorLayer >( QgsMapLayerType::VectorLayer, vectorLayerPath, name, !providerKey.isEmpty() ? providerKey : QLatin1String( "ogr" ), true );
 }
 
 void QgisApp::addMapLayer( QgsMapLayer *mapLayer )
@@ -16414,103 +15587,7 @@ void QgisApp::renameView()
 
 QgsRasterLayer *QgisApp::addRasterLayer( QString const &uri, QString const &baseName, QString const &providerKey )
 {
-  return addLayerPrivate< QgsRasterLayer >( QgsMapLayerType::RasterLayer, uri, baseName, !providerKey.isEmpty() ? providerKey : QLatin1String( "gdal" ), true );
-}
-
-bool QgisApp::addRasterLayers( QStringList const &files, bool guiWarning )
-{
-  if ( files.empty() )
-  {
-    return false;
-  }
-
-  QgsCanvasRefreshBlocker refreshBlocker;
-
-  // this is messy since some files in the list may be rasters and others may
-  // be ogr layers. We'll set returnValue to false if one or more layers fail
-  // to load.
-  bool returnValue = true;
-  for ( const QString &src : files )
-  {
-    QString errMsg;
-    bool ok = false;
-
-    // if needed prompt for zipitem layers
-    QString vsiPrefix = QgsZipItem::vsiPrefix( src );
-    if ( ( !src.startsWith( QLatin1String( "/vsi" ), Qt::CaseInsensitive ) || src.endsWith( QLatin1String( ".zip" ) ) || src.endsWith( QLatin1String( ".tar" ) ) ) &&
-         ( vsiPrefix == QLatin1String( "/vsizip/" ) || vsiPrefix == QLatin1String( "/vsitar/" ) ) )
-    {
-      if ( askUserForZipItemLayers( src, { QgsMapLayerType::RasterLayer } ) )
-        continue;
-    }
-
-    const bool isVsiCurl { src.startsWith( QLatin1String( "/vsicurl" ), Qt::CaseInsensitive ) };
-    const bool isRemoteUrl { src.startsWith( QLatin1String( "http" ) ) || src == QLatin1String( "ftp" ) };
-
-    std::unique_ptr< QgsTemporaryCursorOverride > cursorOverride;
-    if ( isVsiCurl || isRemoteUrl )
-    {
-      cursorOverride = std::make_unique< QgsTemporaryCursorOverride >( Qt::WaitCursor );
-      visibleMessageBar()->pushInfo( tr( "Remote layer" ), tr( "loading %1, please wait …" ).arg( src ) );
-      qApp->processEvents();
-    }
-
-    if ( QgsRasterLayer::isValidRasterFileName( src, errMsg ) )
-    {
-      QFileInfo myFileInfo( src );
-
-      // set the layer name to the file base name unless provided explicitly
-      QString layerName;
-      const QVariantMap uriDetails = QgsProviderRegistry::instance()->decodeUri( QStringLiteral( "gdal" ), src );
-      if ( !uriDetails[ QStringLiteral( "layerName" ) ].toString().isEmpty() )
-      {
-        layerName = uriDetails[ QStringLiteral( "layerName" ) ].toString();
-      }
-      else
-      {
-        layerName = QgsProviderUtils::suggestLayerNameFromFilePath( src );
-      }
-
-      // try to create the layer
-      cursorOverride.reset();
-      QgsRasterLayer *layer = addLayerPrivate< QgsRasterLayer >( QgsMapLayerType::RasterLayer, src, layerName, QStringLiteral( "gdal" ), guiWarning );
-
-      if ( layer && layer->isValid() )
-      {
-        //only allow one copy of a ai grid file to be loaded at a
-        //time to prevent the user selecting all adfs in 1 dir which
-        //actually represent 1 coverage,
-
-        if ( myFileInfo.fileName().endsWith( QLatin1String( ".adf" ), Qt::CaseInsensitive ) )
-        {
-          break;
-        }
-      }
-      // if layer is invalid addLayerPrivate() will show the error
-
-    } // valid raster filename
-    else
-    {
-      ok = false;
-
-      // Issue message box warning unless we are loading from cmd line since
-      // non-rasters are passed to this function first and then successfully
-      // loaded afterwards (see main.cpp)
-      if ( guiWarning )
-      {
-        QString msg = tr( "%1 is not a supported raster data source" ).arg( src );
-        if ( !errMsg.isEmpty() )
-          msg += '\n' + errMsg;
-
-        visibleMessageBar()->pushMessage( tr( "Unsupported Data Source" ), msg, Qgis::MessageLevel::Critical );
-      }
-    }
-    if ( ! ok )
-    {
-      returnValue = false;
-    }
-  }
-  return returnValue;
+  return QgsAppLayerHandling::addRasterLayer( uri, baseName, providerKey );
 }
 
 QgsPluginLayer *QgisApp::addPluginLayer( const QString &uri, const QString &baseName, const QString &providerKey )
